requirements.txt
pytorch-lightning==1.4.5
torchmetrics==0.5.0
scipy==1.5.4
numpy==1.19.5
tqdm
pytorch-fid==0.2.0
pandas==1.1.5
lpips==0.1.4
lmdb==1.2.1
ftfy
regex